Move the AdhocServerScreen to its own file, remove the "Toggle List" button
