+++ assets/lang/sv_SE.ini
@@ -497,8 +497,8 @@ Supported = Stöds
 There is no data = Det finns ingen data.
 This change will not take effect until PPSSPP is restarted. = Denna ändring träder inte i kraft förrän PPSSPP har startats om.
 This will overwrite the existing configuration = Detta kommer att skriva över den befintliga konfigurationen # AI translated
-Toggle All = Vänd alla
-Toggle List = Vänd lista
+Toggle All = Toggla alla
+Toggle List = Toggla lista
 Top Center = Överkant centrerad
 Top Left = Överkant vänster
 Top Right = Överkant höger
